refactor:organizar DAG separando lógica de scraping , banco de dados e utils em módulos
